test: ✅ add comprehensive integration tests for database infrastructure and agent runs
Add 13 comprehensive test modules covering critical backend functionality:

**Database Infrastructure Tests (T1.1):**
- Pool creation, WAL mode, foreign key enforcement
- Full table schema validation (workspaces, threads, messages, runs, tool_calls)
- Critical index verification
- AppError serialization testing

**Agent Run Lifecycle Tests (T1.5):**
- Complete run state machine (created → dispatching → running ↔ waiting_approval → terminal)
- Crash recovery marking dangling runs as interrupted
- Active runs partial indexing (non-terminal only)
- One active run per thread constraint
- Effective model plan persistence
- Runtime session spec builder with profile resolution, plan mode guardrails,
  thinking level support, and structured workspace context

**New Test Coverage:**
- 948-line agent_run.rs with full state transition coverage
- 450-line database_infrastructure.rs with schema/index validation
- tool_gateway.rs (888 lines) for tool execution testing
- test_helpers module for consistent test data seeding

Ensures robust backend reliability and comprehensive test coverage for core agent/thread/run functionality.
